requirements.txt
streamlit
pandas
numpy
json
plotly.express
os